q1/notebook-llm-multimodal-rag/requirements.txt
fastapi
uvicorn
httpx
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-ollama
pydantic
chromadb
langchain-docling
ollama
python-multipart
